requirements.txt
python-fasthtml==0.6.9
uvicorn==0.31.0
markdown==3.7
